requirements.txt
pytorch1.0.0
numpy
python2